split fetch-by-code handler
