+++ apps/asap-server/src/handlers/webhooks/fetch-by-code/handler.ts
@@ -1,0 +1,5 @@
+import Users from '../../../controllers/users';
+import { Handler } from '../../../utils/types';
+import { fetchUserByCodeHandlerFactory } from './fetch-by-code';
+
+export const handler: Handler = fetchUserByCodeHandlerFactory(new Users());
