--- apps/asap-server/src/handlers/webhooks/webhook-fetch-by-code.ts
@@ -1,32 +1,0 @@
-import Joi from '@hapi/joi';
-import { framework as lambda } from '@asap-hub/services-common';
-import { http } from '../../utils/instrumented-framework';
-
-import Users from '../../controllers/users';
-import { Handler } from '../../utils/types';
-import validateRequest from '../../utils/validate-auth0-request';
-
-export const handler: Handler = http(
-  async (request: lambda.Request): Promise<lambda.Response> => {
-    await validateRequest(request);
-
-    const paramsSchema = Joi.object({
-      code: Joi.string().required(),
-    }).required();
-
-    const { code } = lambda.validate(
-      'params',
-      request.params,
-      paramsSchema,
-    ) as {
-      code: string;
-    };
-
-    const users = new Users(request.headers);
-    const user = await users.fetchByCode(code);
-
-    return {
-      payload: user,
-    };
-  },
-);
